requirements.txt
flask==2.2.3
flask-cors==3.0.10
PyYAML==6.0
rich==13.3.2
requests==2.28.2
gunicorn==20.1.0
python-dotenv==1.0.0 